+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Black screen after suspend or udev rule reload
 - IPC socket sticking around after SIGTERM/SIGINT
 - Default configuration failing without all used commands available
+- Broken textures after DRM device reset
 
 ## 1.4.0 - 2026-03-19